+++ cmd/vice/launchcontrol.go
@@ -279,10 +279,12 @@ func (lc *LaunchControlWindow) cleanupAllAircraft() {
 	lc.cleanupOverflights()
 }
 
-func (lc *LaunchControlWindow) Draw(eventStream *sim.EventStream, p platform.Platform) {
+func (lc *LaunchControlWindow) Draw(eventStream *sim.EventStream, p platform.Platform, config *Config) {
 	showLaunchControls := true
 	imgui.SetNextWindowSizeConstraints(imgui.Vec2{300, 100}, imgui.Vec2{-1, float32(p.WindowSize()[1]) * 19 / 20})
+	applyPinWindowClass("Launch Control", config)
 	imgui.BeginV("Launch Control", &showLaunchControls, imgui.WindowFlagsAlwaysAutoResize)
+	drawPinButton("Launch Control", config)
 
 	ctrl := lc.client.State.LaunchConfig.Controller